Sprint 11 hw. fix2
