+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -16,6 +16,30 @@ import java.util.*;
 public class FilmService {
     private final InMemoryFilmStorage inMemoryFilmStorage;
     private final InMemoryUserStorage inMemoryUserStorage;
+    private final Set<Film> topChartedFilms = new TreeSet<>(comparator);
+
+    private static final Comparator<Film> comparator = (film1, film2) -> {
+        return film2.getLikes().size() - film1.getLikes().size();
+    };
+
+    public Film modifyFilm(Film film) {
+        Film filmChecked = inMemoryFilmStorage.findById(film.getId()).orElseThrow(() ->
+                new NotFoundException("Film's id doesn't in database"));
+
+        delFromTopChart(filmChecked);
+        addToTopChart(film);
+
+        return inMemoryFilmStorage.updateFilm(film);
+    }
+
+    public Film removeFilm(Integer id) {
+        Film filmChecked = inMemoryFilmStorage.findById(id).orElseThrow(() ->
+                new NotFoundException("Film's id doesn't in database"));
+
+        delFromTopChart(filmChecked);
+
+        return inMemoryFilmStorage.deleteFilm(id);
+    }
 
     public Film likeFilm(Integer id, Integer userId) {
         Film film = inMemoryFilmStorage.findById(id).orElseThrow(() ->
@@ -24,7 +48,9 @@ public class FilmService {
                 new NotFoundException("User's id doesn't in database"));
         log.trace("The film and the user are in database. Start of adding like...");
 
+        delFromTopChart(film);
         film.addLike(userId);
+        addToTopChart(film);
         inMemoryFilmStorage.updateFilm(film);
         log.trace("Like from user with id: {} had been put", userId);
         return film;
@@ -37,14 +63,32 @@ public class FilmService {
                 new NotFoundException("User's id doesn't in database"));
         log.trace("The film and the user are in database. Start of deleting like...");
 
+        delFromTopChart(film);
         film.delLike(userId);
+        addToTopChart(film);
         inMemoryFilmStorage.updateFilm(film);
         log.trace("Like from user with id: {} had been deleted", userId);
         return film;
     }
 
-    public List<Film> getTopChart(int count) {
+    private void addToTopChart(Film film) {
+        if (!film.getLikes().isEmpty()) {
+            topChartedFilms.add(film);
+            log.debug("Film was successfully added to topChartedFilms");
+        }
+    }
+
+    private void delFromTopChart(Film film) {
+        topChartedFilms.remove(film);
+        log.debug("Film was successfully deleted from topChartedFilms");
+    }
+
+    public List<Film> getTopChart(Integer count) {
         log.trace("Getting topChart in progress");
-        return inMemoryFilmStorage.getTopChart(count);
+        if (topChartedFilms.size() < count) {
+            return new ArrayList<>(topChartedFilms);
+        } else {
+            return new ArrayList<>(topChartedFilms).subList(0, count);
+        }
     }
 }
