+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -3,11 +3,9 @@ package ru.yandex.practicum.filmorate.storage.user;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Slf4j
 @Component
@@ -36,26 +34,10 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public User deleteUser(Integer id) {
-        if (users.containsKey(id)) {
-            User user = users.remove(id);
-            log.debug("User: {} was deleted", user.getName());
+        User user = users.remove(id);
+        log.debug("User: {} was deleted", user.getName());
 
-            List<Integer> friends = new ArrayList<>(user.getFriends());
-
-            if (friends.isEmpty()) {
-                return user;
-            }
-
-            friends.stream()
-                    .map(friendId -> findById(friendId).get())
-                    .peek(user1 -> user1.delFriend(id))
-                    .collect(Collectors.toList());
-
-            return user;
-        } else {
-            log.warn("User with id: {} was not deleted as he/she hadn't been added before", id);
-            throw new ValidationException("Trying to delete user, that hadn't been added before");
-        }
+        return user;
     }
 
     @Override
@@ -65,33 +47,13 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public Optional<User> findById(Integer id) {
-        return users.entrySet()
-                .stream()
-                .filter(entry -> entry.getKey().equals(id))
-                .map(Map.Entry::getValue)
-                .findFirst();
+        return Optional.ofNullable(users.get(id));
     }
 
     @Override
-    public List<User> findFriends(User user) {
-        return user.getFriends()
+    public List<User> findByIds(List<Integer> usersId) {
+        return usersId
                 .stream()
-                .map(this::findById)
-                .map(Optional::get)
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public List<User> findCommonFriends(User user, User otherUser) {
-        List<Integer> commonFriendsId = user.getFriends().stream()
-                .filter(otherUser.getFriends()::contains)
-                .toList();
-
-        if (commonFriendsId.isEmpty()) {
-            return List.of();
-        }
-
-        return commonFriendsId.stream()
                 .map(this::findById)
                 .map(Optional::get)
                 .toList();
